feature(rework-bugs): Adapt collection immutable object.
