+++ backend/src/main/java/org/eclipse/tractusx/traceability/investigations/adapters/rest/InvestigationsController.java
@@ -192,7 +192,7 @@ public class InvestigationsController {
 	public void updateInvestigation(@PathVariable Long investigationId, @Valid @RequestBody UpdateInvestigationRequest updateInvestigationRequest) {
 		validate(updateInvestigationRequest);
 		logger.info(API_LOG_START + "/{}/update with params {}", investigationId, updateInvestigationRequest);
-		investigationsReceiverService.updateInvestigationPublisher(traceabilityProperties.getBpn(), investigationId, updateInvestigationRequest.status(), updateInvestigationRequest.reason());
+		investigationsPublisherService.updateInvestigationPublisher(traceabilityProperties.getBpn(), investigationId, updateInvestigationRequest.status(), updateInvestigationRequest.reason());
 	}
 }
