+++ backend/src/test/java/org/eclipse/tractusx/traceability/investigations/domain/service/InvestigationsPublisherServiceTest.java
@@ -2,29 +2,32 @@ package org.eclipse.tractusx.traceability.investigations.domain.service;
 
 import org.eclipse.tractusx.traceability.assets.domain.ports.AssetRepository;
 import org.eclipse.tractusx.traceability.common.model.BPN;
-import org.eclipse.tractusx.traceability.investigations.domain.model.Investigation;
-import org.eclipse.tractusx.traceability.investigations.domain.model.InvestigationId;
-import org.eclipse.tractusx.traceability.investigations.domain.model.InvestigationStatus;
-import org.eclipse.tractusx.traceability.investigations.domain.model.Severity;
+import org.eclipse.tractusx.traceability.investigations.domain.model.*;
+import org.eclipse.tractusx.traceability.investigations.domain.model.exception.InvestigationReceiverBpnMismatchException;
+import org.eclipse.tractusx.traceability.investigations.domain.model.exception.InvestigationStatusTransitionNotAllowed;
 import org.eclipse.tractusx.traceability.investigations.domain.ports.InvestigationsRepository;
 import org.eclipse.tractusx.traceability.testdata.AssetTestDataFactory;
 import org.eclipse.tractusx.traceability.testdata.InvestigationTestDataFactory;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.time.Clock;
 import java.time.Instant;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyBoolean;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.never;
 
 @ExtendWith(MockitoExtension.class)
 class InvestigationsPublisherServiceTest {
@@ -116,6 +119,144 @@ class InvestigationsPublisherServiceTest {
 		verify(investigationsReadService).loadInvestigation(investigationId);
 		verify(repository).update(investigation);
 		verify(notificationsService).updateAsync(any(), anyBoolean());
+	}
+
+	@Test
+	@DisplayName("Test updateInvestigation is valid")
+	void testUpdateInvestigation() {
+
+		// Given
+		BPN bpn = BPN.of("recipientBPN");
+		Long investigationIdRaw = 1L;
+		InvestigationStatus status = InvestigationStatus.ACKNOWLEDGED;
+		String reason = "the update reason";
+
+		List<AffectedPart> affectedParts = List.of(new AffectedPart("partId"));
+		Notification notification = new Notification(
+			"123",
+			"id123",
+			"senderBPN",
+			"recipientBPN",
+			"senderAddress",
+			"agreement",
+			"information",
+			InvestigationStatus.RECEIVED,
+			affectedParts,
+			Instant.now(),
+			Severity.MINOR
+		);
+
+		Notification notification2 = new Notification(
+			"456",
+			"id123",
+			"senderBPN",
+			"recipientBPN",
+			"senderAddress",
+			"agreement",
+			"information",
+			InvestigationStatus.RECEIVED,
+			affectedParts,
+			Instant.now(),
+			Severity.MINOR
+		);
+		List<Notification> notifications = new ArrayList<>();
+		notifications.add(notification);
+		notifications.add(notification2);
+
+		Investigation investigationTestData = InvestigationTestDataFactory.createInvestigationTestDataWithNotificationList(InvestigationStatus.RECEIVED, "recipientBPN", notifications);
+
+		when(investigationsReadService.loadInvestigation(any(InvestigationId.class))).thenReturn(investigationTestData);
+
+		// When
+		investigationsPublisherService.updateInvestigationPublisher(bpn, investigationIdRaw, status, reason);
+
+		// Then
+		Mockito.verify(repository).update(investigationTestData);
+		Mockito.verify(notificationsService, times(2)).updateAsync(any(Notification.class), anyBoolean());
+	}
+
+	@Test
+	@DisplayName("Test updateInvestigation is invalid because investigation status transition not allowed")
+	void testUpdateInvestigationInvalid() {
+
+		// Given
+		BPN bpn = BPN.of("recipientBPN");
+		Long investigationIdRaw = 1L;
+		InvestigationStatus status = InvestigationStatus.ACCEPTED;
+		String reason = "the update reason";
+
+		List<AffectedPart> affectedParts = List.of(new AffectedPart("partId"));
+		Notification notification = new Notification(
+			"123",
+			"id123",
+			"senderBPN",
+			"recipientBPN",
+			"senderAddress",
+			"agreement",
+			"information",
+			InvestigationStatus.CREATED,
+			affectedParts,
+			Instant.now(),
+			Severity.MINOR
+		);
+
+		List<Notification> notifications = new ArrayList<>();
+		notifications.add(notification);
+
+		Investigation investigationTestData = InvestigationTestDataFactory.createInvestigationTestDataWithNotificationList(InvestigationStatus.SENT, "recipientBPN", notifications);
+
+		when(investigationsReadService.loadInvestigation(any(InvestigationId.class))).thenReturn(investigationTestData);
+
+		// When
+		assertThrows(InvestigationStatusTransitionNotAllowed.class, () -> {
+			investigationsPublisherService.updateInvestigationPublisher(bpn, investigationIdRaw, status, reason);
+		});
+
+		// Then
+		Mockito.verify(repository, never()).update(investigationTestData);
+		Mockito.verify(notificationsService, never()).updateAsync(any(Notification.class));
+	}
+
+	@Test
+	@DisplayName("Test updateInvestigation is invalid because notification receiverbpn is not same as application bpn")
+	void testUpdateInvestigationInvalidBpnMismatch() {
+
+		// Given
+		BPN bpn = BPN.of("applicationBPN");
+		Long investigationIdRaw = 1L;
+		InvestigationStatus status = InvestigationStatus.ACKNOWLEDGED;
+		String reason = "the update reason";
+
+		List<AffectedPart> affectedParts = List.of(new AffectedPart("partId"));
+		Notification notification = new Notification(
+			"123",
+			"id123",
+			"senderBPN",
+			"recipientBPN",
+			"senderAddress",
+			"agreement",
+			"information",
+			InvestigationStatus.CREATED,
+			affectedParts,
+			Instant.now(),
+			Severity.MINOR
+		);
+
+		List<Notification> notifications = new ArrayList<>();
+		notifications.add(notification);
+
+		Investigation investigationTestData = InvestigationTestDataFactory.createInvestigationTestDataWithNotificationList(InvestigationStatus.RECEIVED, "recipientBPN", notifications);
+
+		when(investigationsReadService.loadInvestigation(any(InvestigationId.class))).thenReturn(investigationTestData);
+
+		// When
+		assertThrows(InvestigationReceiverBpnMismatchException.class, () -> {
+			investigationsPublisherService.updateInvestigationPublisher(bpn, investigationIdRaw, status, reason);
+		});
+
+		// Then
+		Mockito.verify(repository, never()).update(investigationTestData);
+		Mockito.verify(notificationsService, never()).updateAsync(any(Notification.class));
 	}
 
 }
