+++ backend/src/test/java/org/eclipse/tractusx/traceability/investigations/domain/service/InvestigationsReceiverServiceTest.java
@@ -38,19 +38,10 @@ class InvestigationsReceiverServiceTest {
 	private InvestigationsRepository mockRepository;
 
 	@Mock
-	private InvestigationsReadService mockReadService;
-
-	@Mock
 	private NotificationMapper mockNotificationMapper;
 
 	@Mock
 	private InvestigationMapper mockInvestigationMapper;
-
-	@Mock
-	private TraceabilityProperties mockTraceabilityProperties;
-
-	@Mock
-	private NotificationsService mockNotificationsService;
 
 	@InjectMocks
 	private InvestigationsReceiverService service;
@@ -88,144 +79,6 @@ class InvestigationsReceiverServiceTest {
 		service.handleNotificationReceiverCallback(edcNotification);
 		// Then
 		Mockito.verify(mockRepository).save(investigationTestData);
-	}
-
-	@Test
-	@DisplayName("Test updateInvestigation is valid")
-	void testUpdateInvestigation() {
-
-		// Given
-		BPN bpn = BPN.of("recipientBPN");
-		Long investigationIdRaw = 1L;
-		InvestigationStatus status = InvestigationStatus.ACKNOWLEDGED;
-		String reason = "the update reason";
-
-		List<AffectedPart> affectedParts = List.of(new AffectedPart("partId"));
-		Notification notification = new Notification(
-			"123",
-			"id123",
-			"senderBPN",
-			"recipientBPN",
-			"senderAddress",
-			"agreement",
-			"information",
-			InvestigationStatus.RECEIVED,
-			affectedParts,
-			Instant.now(),
-			Severity.MINOR
-		);
-
-		Notification notification2 = new Notification(
-			"456",
-			"id123",
-			"senderBPN",
-			"recipientBPN",
-			"senderAddress",
-			"agreement",
-			"information",
-			InvestigationStatus.RECEIVED,
-			affectedParts,
-			Instant.now(),
-			Severity.MINOR
-		);
-		List<Notification> notifications = new ArrayList<>();
-		notifications.add(notification);
-		notifications.add(notification2);
-
-		Investigation investigationTestData = InvestigationTestDataFactory.createInvestigationTestDataWithNotificationList(InvestigationStatus.RECEIVED, "recipientBPN", notifications);
-
-		when(mockReadService.loadInvestigation(any(InvestigationId.class))).thenReturn(investigationTestData);
-
-		// When
-		service.updateInvestigationPublisher(bpn, investigationIdRaw, status, reason);
-
-		// Then
-		Mockito.verify(mockRepository).update(investigationTestData);
-		Mockito.verify(mockNotificationsService, times(2)).updateAsync(any(Notification.class), anyBoolean());
-	}
-
-	@Test
-	@DisplayName("Test updateInvestigation is invalid because investigation status transition not allowed")
-	void testUpdateInvestigationInvalid() {
-
-		// Given
-		BPN bpn = BPN.of("recipientBPN");
-		Long investigationIdRaw = 1L;
-		InvestigationStatus status = InvestigationStatus.ACCEPTED;
-		String reason = "the update reason";
-
-		List<AffectedPart> affectedParts = List.of(new AffectedPart("partId"));
-		Notification notification = new Notification(
-			"123",
-			"id123",
-			"senderBPN",
-			"recipientBPN",
-			"senderAddress",
-			"agreement",
-			"information",
-			InvestigationStatus.CREATED,
-			affectedParts,
-			Instant.now(),
-			Severity.MINOR
-		);
-
-		List<Notification> notifications = new ArrayList<>();
-		notifications.add(notification);
-
-		Investigation investigationTestData = InvestigationTestDataFactory.createInvestigationTestDataWithNotificationList(InvestigationStatus.SENT, "recipientBPN", notifications);
-
-		when(mockReadService.loadInvestigation(any(InvestigationId.class))).thenReturn(investigationTestData);
-
-		// When
-		assertThrows(InvestigationStatusTransitionNotAllowed.class, () -> {
-			service.updateInvestigationPublisher(bpn, investigationIdRaw, status, reason);
-		});
-
-		// Then
-		Mockito.verify(mockRepository, never()).update(investigationTestData);
-		Mockito.verify(mockNotificationsService, never()).updateAsync(any(Notification.class));
-	}
-
-	@Test
-	@DisplayName("Test updateInvestigation is invalid because notification receiverbpn is not same as application bpn")
-	void testUpdateInvestigationInvalidBpnMismatch() {
-
-		// Given
-		BPN bpn = BPN.of("applicationBPN");
-		Long investigationIdRaw = 1L;
-		InvestigationStatus status = InvestigationStatus.ACKNOWLEDGED;
-		String reason = "the update reason";
-
-		List<AffectedPart> affectedParts = List.of(new AffectedPart("partId"));
-		Notification notification = new Notification(
-			"123",
-			"id123",
-			"senderBPN",
-			"recipientBPN",
-			"senderAddress",
-			"agreement",
-			"information",
-			InvestigationStatus.CREATED,
-			affectedParts,
-			Instant.now(),
-			Severity.MINOR
-		);
-
-		List<Notification> notifications = new ArrayList<>();
-		notifications.add(notification);
-
-		Investigation investigationTestData = InvestigationTestDataFactory.createInvestigationTestDataWithNotificationList(InvestigationStatus.RECEIVED, "recipientBPN", notifications);
-
-		when(mockReadService.loadInvestigation(any(InvestigationId.class))).thenReturn(investigationTestData);
-
-		// When
-		assertThrows(InvestigationReceiverBpnMismatchException.class, () -> {
-			service.updateInvestigationPublisher(bpn, investigationIdRaw, status, reason);
-		});
-
-		// Then
-		Mockito.verify(mockRepository, never()).update(investigationTestData);
-		Mockito.verify(mockNotificationsService, never()).updateAsync(any(Notification.class));
 	}
 
 }
